requirements.txt
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.8.1
django-filter==2.2.0
django-jquery==3.1.0
django-json-schema==0.1.0
django-jsonfield==1.3.1
django-jsonfield-schema==0.3.0
django-jsonfield-widget==0.1.0
django-jsonforms==1.1.0
django-rest-framework==0.1.0
django-user-defined-fields==0.0.4
djangorestframework==3.10.3
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
jsonschema==3.2.0
Markdown==3.1.1
more-itertools==7.2.0
Pillow==6.2.1
plantuml==0.3.0
psycopg2-binary==2.8.4
pyrsistent==0.15.5
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.0.1
zipp==0.6.0
psycopg2==2.7.2